requirements.txt
python-dotenv==1.0.1
Pyrogram==2.0.106
uvloop==0.17.0
tgcrypto==1.2.5
aiogram==3.13.1
aiosqlite==0.19.0
asyncpg==0.28.0
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
outline-vpn-api-async==0.1.1
pyxui-async==1.0.2
SQLAlchemy==2.0.23
dogpile.cache==1.3.3
aiolavapy==0.1.2
yookassa-async==0.1.5
YooMoney-async==0.0.6
cryptomus==1.1
aiocryptopay==0.4.5
WalletPay==1.3.1
